+++ android/src/org/coolreader/sync2/BookmarksContentHandler.java
@@ -164,7 +164,7 @@ class BookmarksContentHandler implements ContentHandler {
 						m_fileInfo.seriesNumber = atoi(m_text);
 						break;
 					case "size":
-						m_fileInfo.size = atoi(m_text);
+						m_fileInfo.size = atol(m_text);
 						break;
 					case "crc32":
 						m_fileInfo.crc32 = atol(m_text);
